refactor: extract agent_roles clusters into private submodules
Decomposes the agent_roles sub-package per
docs/guides/decomposition-pattern.md (#3543, following the #3312
pattern and the plan_parser worked reference):

- _core.py            enums + dataclasses + shared reviewer block list
- _refine_roles.py    refine-phase producers and reviewers
- _plan_roles.py      plan-phase producers/reviewer + applier (#1557)
- _implement_roles.py implement-phase producers and reviewers
- _oversight_roles.py overseer + on-demand utility roles
- _registry.py        AGENT_ROLES registry, phase maps, query helpers

__init__.py is the stable re-export barrel with an explicit __all__;
it keeps every module global the single-file module exposed, including
the externally-referenced _PHASE_ROLES / _PHASE_REVIEWERS /
_REVIEWER_BLOCKED_WRITE and the Role / match_pattern imports.

Pure refactor: every symbol is AST-identical to the pre-split file
except the relative-import depth of dependency_graph inside
detect_write_overlaps. Largest submodule is 452 lines, so the
allowlist entry is removed; the shared/CLAUDE.md seam table gains the
agent_roles row.

Closes #3543
