+++ scripts/file-size-allowlist.yaml
@@ -28,5 +28,3 @@ caps:
 files:
   orchestrator/concurrent_executor.py:
     issue: "3498"
-  shared/egg_contracts/agent_roles.py:
-    issue: "3543"
